Delete footnotes.js file and transfer logic over to general.js
